refactor(cli): move board rows + markers to tui_board_rows.go (PRD #1009 M3)
